pyproject.toml
[tool.poetry]
name = "django-blog-project"
version = "0.9.0"
description = ""
authors = ["zmrenwu <zmrenwu@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6, <4"
django = "^3.1.1"
django-allauth = "^0.42.0"
django-contrib-comments = "^1.9.2"
django-mptt = "^0.11.0"
django-notifications-hq = "^1.6.0"
django-haystack = "v3.0b2"
django-model-utils = "^4.0.0"
django-braces = "^1.14.0"
django-imagekit = "^4.0.2"
django-anymail = "^7.2.1"
django-environ = "^0.4.5"
django-extensions = "^3.0.8"
django-webpack-loader = "^0.7.0"
djangorestframework = "^3.11.1"
django-pure-pagination = "^0.3.0"
pytz = "^2020.1"
markdown = "^3.2.2"
beautifulsoup4 = "^4.9.1"
pillow = "^7.2.0"
whoosh = "^2.7.4"
jieba = "^0.42.1"
gunicorn = "^20.0.4"
pymdown-extensions = "^8.0"
django-constance = { version = "^2.7.0", extras = ["database"] }
django-admin-sortable2 = "^0.7.7"
django-celery-email = "^3.0.0"
django-ratelimit = "^3.0.1"
django-taggit = "^1.3.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.2"
flake8 = "^3.8.3"
selenium = "^3.141.0"
fabric = "^2.5.0"
pytest = "^6.0.1"
pytest-django = "^3.9.0"
pytest-cov = "^2.10.1"
freezegun = "^1.0.0"
django-silk = "^4.1.0"
factory-boy = "^3.0.1"
django-test-plus = "^1.4.0"
pytest-sugar = "^0.9.4"
mkdocs = "^1.1.2"
"tl.testing" = "^0.5"
flake8-isort = "^4.0.0"
black = "^20.8b1"
coverage = { version = "^5.3", extras = ["toml"] }

[tool.pytest.ini_options]
minversion = "6.0"
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = "tests.py test_*.py"
norecursedirs = "frontend"
addopts = "--reuse-db --cov=blogproject --cov-report html"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
skip = ["migrations"]

[tool.coverage.run]
include = "blogproject/*"
omit = ["*migrations*", "*tests*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )/
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
